+++ src/ui/index.tsx
@@ -1,4 +1,5 @@
 import ExportDialog from './ExportDialog';
+import ImportDialog from './ImportDialog';
 import ExportSettingTab from './SettingTab';
 
-export { ExportDialog, ExportSettingTab };
+export { ExportDialog, ImportDialog, ExportSettingTab };